# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reseller/clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /reseller/clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /reseller/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - in: query
          name: clientId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /reseller/jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /reseller/items:
    get:
      operationId: listItems
      tags: [items]
      summary: List inventory items
      parameters:
        - in: query
          name: jobId
          required: false
          schema:
            type: integer
        - in: query
          name: clientId
          required: false
          schema:
            type: integer
        - in: query
          name: platform
          required: false
          schema:
            type: string
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /reseller/items/{itemId}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get an inventory item
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    patch:
      operationId: updateItem
      tags: [items]
      summary: Update an inventory item
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an inventory item
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /reseller/ai/analyze-batch:
    post:
      operationId: analyzePhotoBatch
      tags: [ai]
      summary: Run the v2 Gemini batch pipeline on a set of photos for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeBatchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeBatchResult"
...
  /reseller/items/{itemId}/publish:
    post:
      operationId: publishItem
      tags: [items]
      summary: Publish (or save as draft) an item to its routed platform
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishItemResult"
        "400":
          description: Bad request (e.g. Local Pickup or missing description)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Platform API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reseller/items/{itemId}/sync-platform-status:
    post:
      operationId: syncItemPlatformStatus
      tags: [items]
      summary: >
        Poll the routed platform API for the current listing state and update the
        master inventory status (Draft / Listed / Sold / Error). For platforms
        without a public API (Poshmark, Chairish, Facebook Marketplace) returns
        a message instructing the operator to update manually.
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatusResult"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Platform API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reseller/ai/listing-description:
    post:
      operationId: generateListingDescription
      tags: [ai]
      summary: Generate an SEO-optimized platform-specific listing description for an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingDescriptionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDescriptionResult"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    Client:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        household:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [id, createdAt, name]
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        household:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [name]
    Job:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        clientId:
          type: integer
        title:
          type: string
        jobType:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        itemCount:
          type: ["integer", "null"]
      required: [id, createdAt, clientId, title, jobType, status]
    JobInput:
      type: object
      properties:
        clientId:
          type: integer
        title:
          type: string
          minLength: 1
        jobType:
          type: string
          default: estate-clearout
        status:
          type: string
          default: active
        notes:
          type: ["string", "null"]
      required: [clientId, title]
    Item:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        jobId:
          type: integer
        clientId:
          type: integer
        brand:
          type: string
        model:
          type: string
        marketPrice:
          type: ["string", "null"]
        floorPrice:
          type: ["string", "null"]
        platform:
          type: string
        shippingLogic:
          type: ["string", "null"]
        status:
          type: string
        disposition:
          type: string
        listingDescription:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        platformListingId:
          type: ["string", "null"]
        platformListingUrl:
          type: ["string", "null"]
        platformPublishError:
          type: ["string", "null"]
      required:
        - id
        - createdAt
        - updatedAt
        - jobId
        - clientId
        - brand
        - model
        - platform
        - status
        - disposition
    ItemUpdate:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        marketPrice:
          type: ["string", "null"]
        floorPrice:
          type: ["string", "null"]
        platform:
          type: string
        shippingLogic:
          type: ["string", "null"]
        status:
          type: string
        disposition:
          type: string
        listingDescription:
          type: ["string", "null"]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AnalyzeBatchBody:
      type: object
      properties:
        jobId:
          type: integer
        photos:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzeBatchPhoto"
          minItems: 1
      required: [jobId, photos]
...
    AnalyzeBatchResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzedItem"
        savedCount:
          type: integer
        duplicateCount:
          type: integer
        donateCount:
          type: integer
      required: [items, savedCount, duplicateCount, donateCount]
    ListingDescriptionBody:
      type: object
      properties:
        itemId:
          type: integer
      required: [itemId]
    ListingDescriptionResult:
      type: object
      properties:
        itemId:
          type: integer
        platform:
          type: string
        title:
          type: string
        description:
          type: string
      required: [itemId, platform, title, description]
    SyncStatusResult:
      type: object
      properties:
        itemId:
          type: integer
        platform:
          type: string
        newStatus:
          type: string
          description: "The updated status written to the inventory: Draft | Listed | Sold | Error"
        message:
          type: string
        apiCalled:
          type: boolean
      required: [itemId, platform, newStatus, message, apiCalled]
    PublishItemResult:
      type: object
      properties:
        itemId:
          type: integer
        platform:
          type: string
        mode:
          type: string
          description: >
            "live" - listing posted and is active on the platform;
            "draft" - a draft was saved via the platform API (e.g. eBay Seller Hub, Etsy Shop Manager);
            "draft_prepared" - platform has no public listing API (Poshmark, Chairish, Facebook Marketplace);
            a local draft reference is stored and copy-ready content + a direct link to the
            platform listing form are returned for manual completion.
        platformListingId:
          type: ["string", "null"]
        platformListingUrl:
          type: ["string", "null"]
        message:
          type: string
        newStatus:
          type: string
      required: [itemId, platform, mode, message, newStatus]